// src/types/firestore.ts
export interface ArticlesData {
    order: number;
    title: string;
    body: string;
}

export interface HomepageData {
    mainLine1: string,
    mainLine2: string,
    mainLine3: string,
    mainLine4: string,
    mainLine5: string,
    mainLine6: string,
    mainLine7: string;
    mainLine8: string;
    image1: string;
    image2: string;
    image3: string;
    image4: string;
}